최종

+++ src/main/java/com/example/todolistt/service/TodolistService.java
@@ -2,6 +2,7 @@ package com.example.todolistt.service;
 
 import com.example.todolistt.dto.TodolistDto;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,12 +11,12 @@ import java.util.stream.Collectors;
 
 import com.example.todolistt.domain.Todolist;
 import com.example.todolistt.repository.TodolistRepository;
-
 @Service
 public class TodolistService {
 
     private final TodolistRepository todolistRepository;
 
+    @Autowired
     public TodolistService(TodolistRepository todolistRepository) {
         this.todolistRepository = todolistRepository;
     }
@@ -35,15 +36,17 @@ public class TodolistService {
     }
 
     public Todolist updateTodolist(Long id, TodolistDto todolistDto) {
-        return todolistRepository.findById(id)
-                .map(existingTodolist -> {
-                    existingTodolist.setContent(todolistDto.getContent());
-                    existingTodolist.setDate(todolistDto.getDate());
-                    return todolistRepository.save(existingTodolist);
-                }).orElseThrow(() -> new EntityNotFoundException("Todolist not found with id " + id));
+        Todolist todolist = todolistRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("해당 데이터 없음"));
+        todolist.setContent(todolistDto.getContent());
+        todolist.setDate(todolistDto.getDate());
+        return todolistRepository.save(todolist);
     }
 
     public void deleteTodolist(Long id) {
+        if (!todolistRepository.existsById(id)) {
+            throw new EntityNotFoundException("해당 데이터 없음");
+        }
         todolistRepository.deleteById(id);
     }
 }
